requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo
pymongo[srv]==3.12
openpyxl
scikit-learn
dill
pyaml

##-e .